Make CallLogBackupAgent capable of restoring from newer OS version.
Adding restoreAnyVersion removes the restriction that the OS version must
be less than or equal to that of the data being restored.  Added a check
in the onRestore method to check the backup file format version number to
ensure that it is not greater than the current supported version; if it
is we will skip restoring the data.

This means that provided the call log backup format has NOT changed,
you can restore an SDK X+1 backup file to an SDK X device.  This provides
a little bit more flexibility for users to restore from newer devices to
older devices and still retain their call log entries.

Flag: Backup and restore is not flagable; this is also destined for next
release.
Test: Added unit test to verify that we cannot restore newer backup file
versions.
Test: Manual testing was performed to ensure that restoring from newer
OS to older OS works providing the backup file format didn't change.
Bug: 317053396

Change-Id: If21ac4f9afc97251e9e635d564e5ffc90c9161f5